docs: fix template for docs markdown, regenerate docs (#172)
This PR fixes generating markdown docs. There was an issue with it last
time we ran releasing GHA:


https://github.com/equinix-labs/ansible-collection-equinix/actions/runs/8174877007/job/22350694817#step:7:162

I had to fix rendering of the return values, and I renegerated the docs.
The rst doc for dynamic inventory changed slightly too, I'm not sure
why, but I checked it looks ok rendered.