ci: add pr-triggered GHA to validate ansible-specdoc documentation (#180)
This PR adds GHA that will run `make docs`. If that succeeds, it's
guaranteed that docs are properly generated and that the format is OK
for ansible-doc, which should render well in ansible-galaxy.

It's triggered in PRs or manually, auth token not needed so also no
authorization.

I think we need to merge this to test properly.