fix: enable `metal_reserved_ip_block` to create VRF IP reservations (#233)
This fixes a couple problems with the parameter definitions for the
reserved IP block module:

1. The `cidr` and `network` parameters, needed for VRF IP reservations,
were not defined in the module spec
2. The `quantity` parameter, which is unsupported for VRF IP
reservations, was required in order to create _any_ IP reservation

This PR updates the reserved IP block module so that it can be used to
create a VRF IP reservation. The docs are also updated to reference the
`cidr` and `network` parameters.

This was broken out of #232.