chore(ci): move Python version into a file (#230)
We were already specifying a Python version for setup-python actions in
this repo, but this moves the Python version declaration to an external
file. This isn't urgent but it does make it easier to keep the Python
version coordinated across CI workflows and it also enables developers
to know what version of Python is recommended for local development. To
come out of draft, this PR needs to include README updates that document
the existence and potential usage of the `.python-version` file.