chore(ci): disable release comments to avoid secondary rate limit (#246)
The release workflow occasionally fails due to GitHub API rate limiting
on the success comments that are posted to every issue and PR that is
included in the release. While it's possible this problem has been
addressed in the latest version of `semantic-release/github`, I still
think it's better to disable those comments and avoid the problem
entirely.